bans implemented
Former-commit-id: 0c1f728ea0362a24db44001b606cc8ae07e94e32

+++ monero-wallet-java/src/main/java/daemon/model/MoneroBan.java
@@ -8,7 +8,7 @@ public class MoneroBan extends MoneroDaemonModel {
   private String host;  // e.g. 192.168.1.100
   private Integer ip;   // integer formatted IP
   private Boolean isBanned;
-  private Integer seconds;
+  private Long seconds;
   
   public String getHost() {
     return host;
@@ -34,11 +34,11 @@ public class MoneroBan extends MoneroDaemonModel {
     this.isBanned = isBanned;
   }
   
-  public Integer getSeconds() {
+  public Long getSeconds() {
     return seconds;
   }
   
-  public void setSeconds(Integer seconds) {
+  public void setSeconds(Long seconds) {
     this.seconds = seconds;
   }
 }
